requirements.txt
# Core AI and API
google-generativeai>=0.8.0
google-genai>=0.4.0

# Audio Processing
pyaudio>=0.2.11
pygame>=2.5.0
soundfile>=0.12.0

# Computer Vision
opencv-python>=4.8.0
ultralytics>=8.0.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
geopandas>=0.14.0
shapely>=2.0.0

# Web and Mapping
folium>=0.15.0
matplotlib>=3.7.0
networkx>=3.1

# GUI
PySimpleGUI>=5.0.0

# Text Processing
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.20.0
arabic-reshaper>=3.0.0
python-bidi>=0.4.2

# RAG System
langchain>=0.1.0
langchain-community>=0.0.20
langchain-google-genai>=0.0.8
faiss-cpu>=1.7.4

# Utilities
python-dotenv>=1.0.0
loguru>=0.7.0
requests>=2.31.0
geopy>=2.4.0

# Development Tools
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0 